DNA-467 Add search/query support to the graph API DNA-468 Add XPath query language support
Added a lot more implementation and unit tests behind the XPath parser, the XPath AST objects, and the translator that converts an XPath AST into a SQL query model. Found and fixed a number of issues in several of the 'dna-graph' classes, including some tweaks to the SQL query odel, the SQL parser, minor improvements to the JodaDateTime class, and some changes to the exception handling in the 'dna-graph' query model and builder.

At this point, the XPath support is pretty good, though still not complete.  It may be good enough to use for a while - until we have more examples.  The unit tests are verifying not only that the XPath expressions can be parsed, but that they're also converted correctly to an expected SQL representation.

The XPath functionality has not yet been integrated into the JCR implementation, since that requires hooking up the LuceneQueryEngine (which still needs some work).  However, I plan to start putting more of the pieces together and focusing on wrapping up the LuceneQueryEngine.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1286 76366958-4244-0410-ad5e-bbfabb93f86b