DNA-457 Add JSR-170 Locking Optional Feature
Committed patch that provides a locking implementation based on tracking the UUIDs of locked nodes.  This implies that any lockable node must also be referenceable.  The patch allows all of the TCK locking tests to pass as well as a few L2 tests which assumed a locking implementation for the repository.  However, the locking approach does not yet provide any way of integrating the lock/unlock capabilities with the underlying repository.  I will leave the defect open until that need can be addressed as well.

Thanks much for the reviews and suggestions!



git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1295 76366958-4244-0410-ad5e-bbfabb93f86b