DNA-518 JPA connector fails to properly clone or update from another workspace
The previous patch fixed the originally declared issue, but exposed a few regressions. The committed patch fixes most of those regressions and corrects a significant design flaw in the original locking implementation (using jcr:uuid to refer to another nodes uuid).

However, this patch does not correct two key defects. JCR imports (session and workspace) that include collisions with existing nodes that share a UUID will fail if the removeExisting flag is set to true. There is also a minor defect with moving locked nodes on this connector.

Since the original defect is addressed, this defect can be resolved and two new defects can be opened for the other regressions.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1316 76366958-4244-0410-ad5e-bbfabb93f86b