DNA-553 Enhanced the file system connector so that it is much easier to define how non-standard properties can be made to appear on nt:folder, nt:file, and nt:resource nodes.  By default, the file system connector behaves as it has in the past: it only uses the standard properties defined by JCR for files, folders and content nodes.
There is a new CustomPropertiesFactory interface that defines several methods for getting and setting custom properties.  Simply implement this interface, and call FileSystemSource.setCustomPropertiesFactory(...) with an instance of your implementation.  (Or, optionally subclass FileSystemSource, set the factory in the subclass' constructor, and inherit all other functionality).

The connector uses this factory whenever it reads 'nt:folder', 'nt:file', or 'nt:resource' nodes, and includes in these nodes the properties returned by the factory. The factory methods take several arguments, including the ExecutionContext (which can be used to obtain the PropertyFactory, value factories used to create property values, and even the MIME type detector), the java.io.File object, the Location (which includes the path), and other information.  Hopefully this is sufficient to construct the necessary custom properties.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1325 76366958-4244-0410-ad5e-bbfabb93f86b