DNA-559 GraphSession cloneImmediate fails to remove existing nodes from cache because of Set ordering variations
GraphSession.cloneImmediate(...) has a parameter that specifies whether any existing nodes should be removed (from the connector), and if this connector request succeeds and did remove those nodes, the method continues by attempting to purge these nodes from the cache. However, the request collects in a Set the Location for each of the removed nodes, so the order of these nodes cannot be guaranteed. And in fact when the cloneImmediate method walks them, the test was purging a parent node followed by a child node. Since purging the parent also purged all children (and other descendants), attempting to purge the child results in an exception.

The fix is relatively simple: check whether a node being purged is below a node that already was purged.  All unit tests now pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1350 76366958-4244-0410-ad5e-bbfabb93f86b