DNA-546 JCR Workspace and Session Imports Can Fail on JPA Connector
Applied patch that makes some preparatory tweaks to the MapRepository code. MapNode is renamed to DefaultMapNode and now implements a new interface named MapNode. Additionally, some bulk mutators are added.

The first part is a change that had been kicked around during the original development of the MapRepository code and, IIRC, was bypassed as one of those "if we need it later, we'll do it" changes. I'm doing it now to enable the creation of a JpaNode for the new JPA connector model. Instead of implementing the MapNode interface by managing state internally (like DefaultMapNode does), the JpaNode will be an adapter over a JPA entity.

The bulk mutators are added to help limit the number of times that the properties for a node need to be reserialized.


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1361 76366958-4244-0410-ad5e-bbfabb93f86b