DNA-546
Committed patch (DNA-546_jpa_connector_refactor.patch) that refactors the existing JPA connector framework to more readily support the new model that I'm working on.

The patch removes the relationship between the model and the request processor and adds a relationship between the model and its connection. To work within the MapRepository framework, I had to be able to pass a reference to the MapRepository into the MapRequestProcessor. This ruled out using the existing JpaConnection. The cleanest way to isolate the reusable pieces of the code seemed to be to rename JpaConnection to BasicJpaConnection, move it into model.basic, and then replace Model.createRequestProcessor(...) with Model.createConnection(JpaSource). This required a few changes to class visibility (but not constructor visibility) and the addition of some accessors to JpaSource.


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1366 76366958-4244-0410-ad5e-bbfabb93f86b