DNA-565 JCR Implementation Should Support Anonymous Access More Easily
Committed patch that adds a JcrRepository.Option for ANONYMOUS_USER_ROLES. This defaults to null (indicating that an anonymous user will not be granted any roles and will not be able to log in). It can be set to a comma-delimited list of role names that will be provided to an anonymous user. A test case has been added at the JCR layer.

The patch also changes the REST server around a bit to accommodate this new access. If the requesting user is not authenticated, the REST code will now attempt to log in anonymously. I've tested this locally and it did work for me, but I couldn't figure out a good way to write test cases for it and still test the behavior when it's not enabled (short of creating a whole new subproject which doesn't seem worth it).

I added this line to the configRepository.xml in the dna-web-jcr-rest project to test the anonymous access:
                <anonymousUserRoles jcr:primaryType="dna:option" value="connect,readonly"/>

This gave all unauthenticated users the ability to connect to the REST server and read-only privileges for all workspaces in all repositories.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1378 76366958-4244-0410-ad5e-bbfabb93f86b