DNA-577
Applied patch that changes the default setting in JpaSource and updates the reference guide. I changed the sub-sections to be of style <sect1> instead of <sect2> to be compliant with the DTD for the enclosing <chapter> tag. This cleans up an error in the Eclipse Docbook editor.

The patch also re-adds a test that was overridden a few months ago when DNA-466 was still open and cleans up some of the existing connector-level tests for the simple model. 

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1408 76366958-4244-0410-ad5e-bbfabb93f86b