DNA-541 Locking Implementation Does Not Support Timeouts
Committed a patch that implements lock expiration based on the algorithm above modified by Randall's comments. The patch includes two tests to confirm that the functionality works, but one is @Ignored because it depends on garbage collector behavior to test the results and the other is @Ignored because it has to sleep for 30 seconds to test the timeout. 

The patch also addresses specific feedback from the JIRA by integrating the ScheduledExecutionService in the JcrEngine into the DnaEngine's lifecycle methods (start, shutdown, awaitTermination) and synchronizing access to JcrRepository.activeSessions

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1414 76366958-4244-0410-ad5e-bbfabb93f86b