DNA-467 Restructured the SearchEngine functionality to simplify it, and to change over to have implementations be extensions.  To specialize, an implementation now simply subclasses the SearchEngine, and this is instantiated directly.
The SearchEngine is also now oriented around a RequestProcessor.  It's possible to obtain a RequestProcessor to issue multiple requests within a single connection, and this will make it significantly easier to reuse within a connector implementation (as well as within the 'dna-repository' module).  Also, the SearchEngine's abstract RequestProcessor implementation provides default implementations for crawling and indexing entire subgraphs to generate the required UpdatePropertiesRequest and CreateNodeRequest stream.  Specializations are responsible for fully-implementing the RequestProcessor to handle all the different kinds of requests, including AccessQueryRequest and FullTextSearchRequest.

Thus, the LuceneSearchEngine can be instantiated directly, as it is now a subclass of SearchEngine.  It also exists in a new 'extensions/dna-search-lucene' project, and continues to use the two-index design that was implemented previously.  This new implementation uses a customized RequestProcessor implementation.  This also should make it easier to process Changes via the request processor's methods.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1418 76366958-4244-0410-ad5e-bbfabb93f86b