DNA-582 DNA-585 Changed the semantics of the RequestProcessor.close().  Previously, this method released all resources _and_ notified the observer of all changes.  The resources always have to be released and thus close() must be called, yet this was always notifying the observer even in the case of change failures.
The close() method now no longer notifies the observer, and a new notifyObserverOfChanges() method has been added to do the notification.  This makes it possible to always close the resources yet optionally do the notification depending upon other conditions.

The RequestProcessor.process(CompositeRequest) method behavior has also changed. Now, if there is an error on any of the change requests (any request that returns 'false' for isReadOnly()), the processing stops and the error is recorded.

The abstract MapRepositorySource has been changed to take advantage of this.  A new MapRepositoryTransaction has been created, and the MapRepository now has a method to return a new transaction.  The MapRepositoryConnection.execute(...) method has been changed to take advantage of this to obtain a transaction, perform the processing, check for errors on change requests, close the processor (to release any resources), and then either commit or rollback the transaction based upon whether the requests had errors and involved changes.  Thus, the new MapRepositoryTransaction encapsulates all of the locking and rollback behavior for specializations of the MapRepository.

As part of DNA-585, the simple model in the JPA connector has been changed to use this transaction behavior and to commit or rollback the JPA transaction appropriately.  The basic model in the JPA connector doesn't extend the MapRepository classes, but it does the same thing (using the JPA transaction directly in the connection's execute method).  Therefore, this change should resolve DNA-585.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1431 76366958-4244-0410-ad5e-bbfabb93f86b