DNA-583 DNA-582 Added a mechanism by which the InMemory connector maintains the original state of nodes the first time they are changed, so that the original state can be restored should the changes be rolled back.  This does add a little extra overhead as nodes are changed, but there's almost no extra overhead to committing the changes (since the original states are thrown out).  When a rollback does occur, there is (obviously) some extra work restoring the state of the changed nodes.
There is also a new TransactionalConnectorTest abstract class in the 'org.jboss.dna.graph.connector.test' package that can be subclassed to create a connector-specific test to verify that a rollback results in no changes to the state.  A new InMemoryConnectorTransactionalTest concrete class was created to test the InMemory connector.

This design was a tradeoff that achieves simplicity and straightforwardness (and encapsulation within the InMemoryRepositorySource code) while adding only modest overhead to the changes.  It is possible that the design could be improved to lower the overhead of each change, but this change does only add less than 1-2% overhead the 'dna-jcr' unit tests (<0.9sec increase on 59 sec tests, which is actually within the variation common between runs).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1432 76366958-4244-0410-ad5e-bbfabb93f86b