DNA-598 Shutting down JcrEngine returns, but leaves idle threads running (preventing shutdown of VM) for 60 seconds
Fixed the problem with the shutdown leaving idle threads and blocking termination of the VM.  After debugging (and using a new NameThreadFactory that creates threads with names beginning with a specified prefix), it's clear now that these extra threads were owned by the Executor used in the FederatedRepositorySource, and this executor was not being shutdown in the FederatedRepositorySource.close() method.  The Executor was creating the threads with a timeout of 60sec, which makes sense.  Fixing this now properly shuts down the Executor, and this fixes the problem.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1451 76366958-4244-0410-ad5e-bbfabb93f86b