DNA-593
Committed patch that updates the documentation to reflect that the connector is no longer read-only (incl. adding the updatesAllowed property to the docs) and renamed the getSupportsUpdates method to be consistent with the setUpdatesAllowed mutator.

The patch also adds references to the custom properties factory, but does not complete the sub-section describing the behavior yet. 

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1464 76366958-4244-0410-ad5e-bbfabb93f86b