DNA-594 The JDBC Metadata Connector documentation in Reference Guide is missing
Applied patch that fixes the link to the JDBC metadata connector page and brings it up to the same standard as the other connector pages in the reference guide.  Also corrects a few typos in the JPA connector page.

The patch doesn't explicitly list the connector as "experimental" or "technology preview" yet because I'm pretty confident that I'll at least be able to test it against MSSQL and Oracle prior to the 0.7 release.  The page does list the drivers that the connector has been tested against.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1465 76366958-4244-0410-ad5e-bbfabb93f86b