DNA-105 Sequence content from delimited text files
Applied a more refined version of the previous DNA-105 patch that includes documentation (incl. the obligatory Reference Guide chapter), integration with the gettingstarted sequencers example, and a maximumLinesToRead property that allows the sequencers to be configured to not sequencer an entire file.

The patch also modifes StreamSequencerAdapter.setConfiguration to allow for properties to be set on StreamSequencers through DnaConfiguration.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1475 76366958-4244-0410-ad5e-bbfabb93f86b