DNA-578 JPA Connector Simple Model Does Not Support "supportsUpdates" Property
Applied patch that requires support for "allowsUpdates" property in MapRepositorySource implementations and adds a default implementation in MapRequestProcessor.  The patch also adds the property to the Infinispan and JBoss Cache connectors and the simple model in the JPA connector.  Finally, the patch adds a test case for the JPA source and updates the documentation accordingly.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1481 76366958-4244-0410-ad5e-bbfabb93f86b