DNA-464 Add use of file system connector to repositories example
Applied a patch that adds a file system connector (for UFOs, which are so strange that they can't be placed in a normal in-memory repository!) and updates the tests and documentation accordingly*  The patch also removes the need to manually log in to each repository, something I found seriously annoying and impossible to use without reading either the manual or the source code.  I can replace the old code if anyone disagrees with this change.

* - I do not have "graffle" installed, so I did not update any of the images in the Repositories Example chapter of the Getting Started Guide

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1494 76366958-4244-0410-ad5e-bbfabb93f86b