DNA-62
Committed first (working) version of this sequencer. The sequencer calls a Reader implementation (a la the MS Office sequencer) that parses a class file into a metadata object. The metadata object is then passed to a (customizable) recorder object that creates the relevant graph entries. The patch integrates the sequencer with the getting started sequencers example to help verify that it works.

The current version has extensive testing of the class file reader implementation, but very little testing of the default recorder (the object that converts the metadata into a series of graph nodes and properties).

For the defect to be resolved properly, I think these things would need to happen (in no particular order):

1. Align the default output for the class sequencer and the Java sequencer so that they are as identical as possible. This will be very helpful for future analyzer writers.
2. Align the metadata objects between the class sequencer and the Java source sequencer. Both are parsing into objects with the same name. It seems like they could probably be moved to share a common interface and then reuse recorder implementations (solving issue 1).  
3. Move the sequencer into the same package and namespace as the java sequencer. This is a nice to have, but the two sequencers are so tightly related that it makes a lot of sense (to me, anyway)
4. Clean up/add Javadoc and add more testing for the default recorder implementation

Steps 2 and 3 may or may not be appropriate, for all of the good reasons that rhauch noted in his comments on this defect.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1512 76366958-4244-0410-ad5e-bbfabb93f86b