DNA-519 SVN Connector integration test runs in a very long time
Applied patch (DNA-519_repackage_svn.patch) that moves everything back into the svn package.  At this point, the SVN connector has now been fully converted to the path repository framework and can start to leverage the caching support provided therein.  The next step is to write a CachePolicy and WorkspaceCache implementation for the SVN connector to improve read performance.  After that, connection pooling can be considered.  

There is also a need to implement all-or-nothing transaction support.  That will be covered in a separate defect.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1525 76366958-4244-0410-ad5e-bbfabb93f86b